backend/requirements.txt
# FastAPI
fastapi
uvicorn[standard]
python-multipart

# Postgres Driver
psycopg2-binary

# SQLAlchemy (для моделей)
sqlalchemy
sqlmodel

# Alembic (для миграций)
alembic

# JWT (для аутентификации)
python-jose[cryptography]
passlib
argon2-cffi